+++ slacktools/block_kit/blocks/header.py
@@ -1,3 +1,8 @@
+from typing import (
+    List,
+    Union,
+)
+
 from slacktools.block_kit.base import BaseElement
 from slacktools.block_kit.elements.display.text import PlainTextElement
 
@@ -17,3 +22,13 @@ class HeaderBlock(BaseElement):
 
         super().__init__(type=self.type)
         self.length_assertion(self.text.text, 'header_text', 10)
+
+
+class PlainTextHeaderBlock(HeaderBlock):
+
+    def __init__(self, text: Union[str, List[str]], block_id: str = None):
+        self.text = PlainTextElement(text=text)
+        if block_id is not None:
+            self.block_id = block_id
+
+        super().__init__(self.text, block_id=block_id)